pyproject.toml
[tool.poetry]
name = "bittercodeme"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
fastapi = "^0.115.2"
sqlalchemy = "^2.0.35"
email-validator = "2.2.0"
bcrypt = "^4.2.0"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

